// src/app/shared/enums/colors.enum.ts
export enum ColorsEnum {
    Yellow = 'Yellow',
    Blue = 'Blue',
    Beige = 'Beige',
    White = 'White',
    Grey = 'Grey',
    Gold = 'Gold',
    Orange = 'Orange',
    Brown = 'Brown',
    Multicolor = 'Multicolor',
    Black = 'Black',
    Purple = 'Purple',
    Green = 'Green',
    Red = 'Red',
    Navy = 'Navy'
}